backend/requirements.txt
websocket-client
requests
websocket-server
python-dotenv